backend_django/requirements.txt
asgiref==3.11.0
Django==6.0
djangorestframework==3.16.1
gunicorn
psycopg2-binary==2.9.11
sqlparse==0.5.4
tzdata==2025.2
django-cors-headers
djangorestframework-simplejwt[crypto]==5.3.1
python-dotenv==1.0.1
drf-spectacular==0.27.2
whitenoise
